style: move back to custom.py
