+++ src/azure-cli/azure/cli/command_modules/network/aaz/latest/__init__.py
@@ -7,5 +7,3 @@
 
 # pylint: skip-file
 # flake8: noqa
-
-from ._clients import AAZNonRetryableClient
